Time: 0 ms (100.00%) | Memory: 8.1 MB (78.96%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int rem;
-        long long x1=x;
+       int x1=x;
         long long rev=0;
         while(x>0){
             rem=x%10;
